+++ tests/phpunit/Schema/Definition/TableDefinitionTest.php
@@ -34,6 +34,7 @@ class TableDefinitionTest extends \PHPUnit_Framework_TestCase {
 				new FieldDefinition( 'h', FieldDefinition::TYPE_TEXT ),
 				new FieldDefinition( 'i', FieldDefinition::TYPE_INTEGER, FieldDefinition::NOT_NULL, 42 ),
 				new FieldDefinition( 'bi', FieldDefinition::TYPE_BIGINT, FieldDefinition::NOT_NULL, 42 ),
+				new FieldDefinition( 'd', FieldDefinition::TYPE_DECIMAL ),
 			)
 		);