+++ tests/phpunit/Schema/ReportingTableBuilderTest.php
@@ -57,6 +57,7 @@ class ReportingTableBuilderTest extends \PHPUnit_Framework_TestCase {
 		$tables[] = new TableDefinition( 'defaultfieldvalues', array(
 			new FieldDefinition( 'intfield', FieldDefinition::TYPE_INTEGER, true, 42 ),
 			new FieldDefinition( 'bigintfield', FieldDefinition::TYPE_BIGINT, true, 42 ),
+			new FieldDefinition( 'decimalfield', FieldDefinition::TYPE_DECIMAL ),
 		) );
 
 		$tables[] = new TableDefinition( 'notnullfields', array(